refactor(core): move postTask to core/util/postTask
